requirements.txt
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
schedule==1.2.2
loguru==0.7.3
